Conclusione e Fix

+++ backend/sportello/src/main/java/it/mediflow/sportello/service/impl/MediciService.java
@@ -1,9 +1,13 @@
 package it.mediflow.sportello.service.impl;
 
 import it.mediflow.sportello.entity.Medici;
+import it.mediflow.sportello.entity.Specializzazione;
+import it.mediflow.sportello.exceptions.ConflictException;
 import it.mediflow.sportello.exceptions.NotFoundException;
 import it.mediflow.sportello.mappers.IMediciMapper;
 import it.mediflow.sportello.repository.MediciRepository;
+import it.mediflow.sportello.repository.PrenotazioniRepository;
+import it.mediflow.sportello.repository.SpecializzazioneRepository;
 import it.mediflow.sportello.repository.specification.SearchSpecification;
 import it.mediflow.sportello.service.IMediciService;
 import it.mediflow.sportello.web.dto.MediciDto;
@@ -25,6 +29,8 @@ import java.util.Optional;
 public class MediciService implements IMediciService {
 
     private final MediciRepository mediciRepository;
+    private final PrenotazioniRepository prenotazioniRepository;
+    private final SpecializzazioneRepository specializzazioneRepository;
     private final IMediciMapper mediciMapper;
 
     @Override
@@ -45,15 +51,29 @@ public class MediciService implements IMediciService {
     @Override
     public List<String> ricercaPerCodiceFiscale(String cf) {
         Pageable page = PageRequest.of(0, 10); // Limito le corrispondenze a 10
-        List<Medici> medici = mediciRepository.findDistinctByCodiceFiscaleLikeIgnoreCase(cf, page);
+        List<Medici> medici = mediciRepository.findDistinctByCodiceFiscaleContainsIgnoreCase(cf, page);
 
         return medici.stream().map(Medici::getCodiceFiscale).toList();
+    }
+
+    @Override
+    public List<MediciDto> dettaglioPerCodiceFiscale(String cf) {
+        Pageable page = PageRequest.of(0, 10); // Limito le corrispondenze a 10
+        List<Medici> medici = mediciRepository.findDistinctByCodiceFiscaleContainsIgnoreCase(cf, page);
+
+        return mediciMapper.toDTOs(medici);
     }
 
     @Override
     @Transactional
     public MediciDto salvaMedico(MediciDto medico) {
         Medici objSaved = mediciMapper.toEntity(medico);
+
+        Specializzazione specializzazione = specializzazioneRepository.findById(
+                objSaved.getSpecializzazione().getId()
+        ).orElseThrow(() -> new IllegalArgumentException("Specializzazione non trovata"));
+
+        objSaved.setSpecializzazione(specializzazione);
         objSaved = mediciRepository.save(objSaved);
 
         return mediciMapper.toDTO(objSaved);
@@ -63,6 +83,9 @@ public class MediciService implements IMediciService {
     @Transactional
     public void eliminaMedico(Long id) {
         if(!mediciRepository.existsById(id)) throw new NotFoundException("Medico non trovato");
+
+        if(prenotazioniRepository.existsByMedico_Id(id))
+            throw new ConflictException("Impossibile eliminare un medico con prenotazioni");
 
         mediciRepository.deleteById(id);
     }
